backend/requirements.txt
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==6.0.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
